+++ Arion/include/Arion/debug/DebugImplementation.hpp
@@ -40,8 +40,8 @@ public:
         intersection::gjk::Simplex&,
         SimpleShape const&,
         SimpleShape const&,
-        glm::dvec3,
-        glm::dvec3)>;
+        glm::vec3,
+        glm::vec3)>;
 
     /**
      * @brief EPA debug call function
@@ -68,8 +68,8 @@ public:
             intersection::gjk::Simplex& simplex,
             SimpleShape const& aShape,
             SimpleShape const& bShape,
-            glm::dvec3 supportVertex,
-            glm::dvec3 direction
+            glm::vec3 supportVertex,
+            glm::vec3 direction
         )
     {
         GetEpaCallback<ConvexHullBuffer>()(
@@ -150,8 +150,8 @@ private:
         arion::intersection::gjk::Simplex&,
         arion::SimpleShape const&,
         arion::SimpleShape const&,
-        glm::dvec3,
-        glm::dvec3)
+        glm::vec3,
+        glm::vec3)
     {
     }
 };
