+++ Arion/sources/Intersection.cpp
@@ -16,65 +16,65 @@ namespace arion
 {
 
 bool intersection::CheckRaySphereIntersection(
-    glm::dvec3 const& raySphere, double sphereRadius, glm::dvec3 const& rayDirection
+    glm::vec3 const& raySphere, float sphereRadius, glm::vec3 const& rayDirection
 )
 {
-    double const tCenter = glm::dot(raySphere, rayDirection);
-    double const distanceSquare = glm::dot(raySphere, raySphere) - tCenter * tCenter;
+    float const tCenter = glm::dot(raySphere, rayDirection);
+    float const distanceSquare = glm::dot(raySphere, raySphere) - tCenter * tCenter;
 
     return sphereRadius * sphereRadius - distanceSquare >= 0.0;
 }
 
 intersection::RayIntersectionFactors
 intersection::CalculateRaySphereIntersectionFactors(
-    glm::dvec3 const& raySphere, double sphereRadius, glm::dvec3 const& rayDirection
+    glm::vec3 const& raySphere, float sphereRadius, glm::vec3 const& rayDirection
 )
 {
-    double const tCenter = glm::dot(raySphere, rayDirection);
-    double const distanceSquare = glm::dot(raySphere, raySphere) - tCenter * tCenter;
-    double const tDelta = glm::sqrt(sphereRadius * sphereRadius - distanceSquare);
+    float const tCenter = glm::dot(raySphere, rayDirection);
+    float const distanceSquare = glm::dot(raySphere, raySphere) - tCenter * tCenter;
+    float const tDelta = glm::sqrt(sphereRadius * sphereRadius - distanceSquare);
 
     return {tCenter - tDelta, tCenter + tDelta};
 }
 
 intersection::RayIntersectionFactors
 intersection::CalculateRayAabbIntersectionFactors(
-    glm::dvec3 const& boxMinPoint, glm::dvec3 const& boxMaxPoint,
-    glm::dvec3 const& rayDirection, glm::dvec3 const& rayOrigin
+    glm::vec3 const& boxMinPoint, glm::vec3 const& boxMaxPoint,
+    glm::vec3 const& rayDirection, glm::vec3 const& rayOrigin
 )
 {
-    double const t1 = (boxMinPoint.x - rayOrigin.x) / rayDirection.x;
-    double const t2 = (boxMaxPoint.x - rayOrigin.x) / rayDirection.x;
-    double const t3 = (boxMinPoint.y - rayOrigin.y) / rayDirection.y;
-    double const t4 = (boxMaxPoint.y - rayOrigin.y) / rayDirection.y;
-    double const t5 = (boxMinPoint.z - rayOrigin.z) / rayDirection.z;
-    double const t6 = (boxMaxPoint.z - rayOrigin.z) / rayDirection.z;
+    float const t1 = (boxMinPoint.x - rayOrigin.x) / rayDirection.x;
+    float const t2 = (boxMaxPoint.x - rayOrigin.x) / rayDirection.x;
+    float const t3 = (boxMinPoint.y - rayOrigin.y) / rayDirection.y;
+    float const t4 = (boxMaxPoint.y - rayOrigin.y) / rayDirection.y;
+    float const t5 = (boxMinPoint.z - rayOrigin.z) / rayDirection.z;
+    float const t6 = (boxMaxPoint.z - rayOrigin.z) / rayDirection.z;
 
-    double const tmin = glm::max(glm::max(glm::min(t1, t2), glm::min(t3, t4)), glm::min(t5, t6));
-    double const tmax = glm::min(glm::min(glm::max(t1, t2), glm::max(t3, t4)), glm::max(t5, t6));
+    float const tmin = glm::max(glm::max(glm::min(t1, t2), glm::min(t3, t4)), glm::min(t5, t6));
+    float const tmax = glm::min(glm::min(glm::max(t1, t2), glm::max(t3, t4)), glm::max(t5, t6));
 
     return {tmin, tmax};
 }
 
 intersection::AabbExtremalVertices
 intersection::MakeExtremalVerticesAabb(
-    glm::dvec3 const& i, glm::dvec3 const& j, glm::dvec3 const& k
+    glm::vec3 const& i, glm::vec3 const& j, glm::vec3 const& k
 )
 {
-    glm::dmat3 const boxModelMatrixInverse = glm::inverse(
-        glm::dmat3{glm::normalize(i), glm::normalize(j), glm::normalize(k)}
+    glm::mat3 const boxModelMatrixInverse = glm::inverse(
+        glm::mat3{glm::normalize(i), glm::normalize(j), glm::normalize(k)}
     );
 
-    glm::dmat3 const boxAxesModelSpace{
+    glm::mat3 const boxAxesModelSpace{
         boxModelMatrixInverse * i, boxModelMatrixInverse * j, boxModelMatrixInverse * k
     };
 
-    auto const findMaxAbs = [](double a, double b)
+    auto const findMaxAbs = [](float a, float b)
     {
         return std::abs(a) < std::abs(b);
     };
 
-    glm::dvec3 const maxVertex = glm::dvec3{
+    glm::vec3 const maxVertex = glm::vec3{
         glm::abs(*std::max_element(
                 glm::value_ptr(boxAxesModelSpace[0]), glm::value_ptr(boxAxesModelSpace[0]) + 3, findMaxAbs)
         ),
@@ -96,10 +96,10 @@ bool intersection::CheckRayIntersectionFactors(RayIntersectionFactors factors)
 }
 
 bool intersection::IsPointInsideTriangle(
-    glm::dvec3 const& triangleVertex1, glm::dvec3 const& triangleVertex2, glm::dvec3 const& triangleVertex3, glm::dvec3 const& point
+    glm::vec3 const& triangleVertex1, glm::vec3 const& triangleVertex2, glm::vec3 const& triangleVertex3, glm::vec3 const& point
 )
 {
-    double const distance = epona::HyperPlane{triangleVertex1, triangleVertex2, triangleVertex3}.Distance(point);
+    float const distance = epona::HyperPlane{triangleVertex1, triangleVertex2, triangleVertex3}.Distance(point);
     if (!epona::fp::IsZero(distance))
     {
         return false;
